+++ elements/layercontrol/src/components/layerList.js
@@ -16,6 +16,7 @@ export class EOxLayerControlLayerList extends LitElement {
     layers: { attribute: false },
     map: { attribute: false, state: true },
     titleProperty: { attribute: "title-property", type: String },
+    showLayerZoomState: { attribute: "show-layer-zoom-state", type: Boolean },
     tools: { attribute: false },
     unstyled: { type: Boolean },
     noShadow: { type: Boolean },
@@ -52,6 +53,11 @@ export class EOxLayerControlLayerList extends LitElement {
      * The layer title property
      */
     this.titleProperty = "title";
+
+    /**
+     * Show layer state based on zoom level or not
+     */
+    this.showLayerZoomState = false;
 
     /**
      * Render the element without additional styles
@@ -125,6 +131,7 @@ export class EOxLayerControlLayerList extends LitElement {
                                 .idProperty=${this.idProperty}
                                 .map=${this.map}
                                 .titleProperty=${this.titleProperty}
+                                .showLayerZoomState=${this.showLayerZoomState}
                                 .tools=${this.tools}
                                 .unstyled=${this.unstyled}
                                 @changed=${() => this.requestUpdate()}
@@ -135,7 +142,9 @@ export class EOxLayerControlLayerList extends LitElement {
                               <eox-layercontrol-layer
                                 .noShadow=${true}
                                 .layer=${layer}
+                                .map=${this.map}
                                 .titleProperty=${this.titleProperty}
+                                .showLayerZoomState=${this.showLayerZoomState}
                                 .tools=${this.tools}
                                 .unstyled=${this.unstyled}
                                 @changed=${() => {