+++ elements/layercontrol/src/helpers.js
@@ -100,3 +100,112 @@ export const getLayerType = (layer, map) => {
     ? "vector"
     : "raster";
 };
+
+/**
+ * Returns whether zoom layer state be enabled or not for a particular layer
+ * @param {import("ol/layer").Layer | import("ol/layer").Group} layer
+ * @param {Boolean} showLayerZoomState
+ * @returns {Boolean} state
+ */
+export const isLayerZoomStateRequired = (layer, showLayerZoomState) => {
+  const minZoom = layer.get("minZoom");
+  const maxZoom = layer.get("maxZoom");
+
+  if (showLayerZoomState && (minZoom !== -Infinity || maxZoom !== Infinity))
+    return true;
+  else return false;
+};
+
+/**
+ * Returns layer visibility state based on minZoom and maxZoon
+ * with respective of current zoom level
+ * @param {import("ol/layer").Layer | import("ol/layer").Group} layer
+ * @param {import("ol").Map} map
+ * @param {Boolean} showLayerZoomState
+ * @returns {Boolean} state
+ */
+export const isLayerVisibleBasedOnZoomState = (
+  layer,
+  map,
+  showLayerZoomState
+) => {
+  if (!layer || !map) return false;
+
+  if (!isLayerZoomStateRequired(layer, showLayerZoomState)) return true;
+
+  const minZoom = layer.get("minZoom");
+  const maxZoom = layer.get("maxZoom");
+  const zoom = map.getView().getZoom();
+
+  return zoom > minZoom && zoom < maxZoom ? true : false;
+};
+
+/**
+ * Update layer's URL using input values
+ * @param {String} url
+ * @param {Object} values
+ * @param {import("ol/layer").Layer | false} layer
+ */
+export function updateUrl(url, values) {
+  const searchParams = new URL(url).searchParams;
+
+  Object.entries(values).forEach(([key, value]) => {
+    if (typeof value === "object" && !Array.isArray(value) && value !== null) {
+      Object.keys(value).forEach((k) => {
+        searchParams.set(k, value[k]);
+      });
+    } else searchParams.set(key, value);
+  });
+
+  const urlWithPath = url.split("?")[0];
+  const searchParamsStr = searchParams.toString();
+  const newUrl = `${urlWithPath}?${searchParamsStr}`;
+
+  return newUrl;
+}
+
+/**
+ * Recursive function to retrieve startVal inside schema
+ *
+ * @param {{[key: string]: any}} schema
+ * @param {{[key: string]: any}} queryParams
+ */
+export function getNestedStartVals(schema, queryParams) {
+  let startVals = {};
+
+  for (const key in schema) {
+    const type = schema[key].type;
+    if (type && type !== "object") {
+      startVals[key] =
+        type === "number" ? Number(queryParams[key]) : queryParams[key];
+    } else if (typeof schema[key] === "object" && schema[key]?.properties) {
+      const nestedStartVals = getNestedStartVals(
+        schema[key].properties,
+        queryParams
+      );
+      if (Object.keys(nestedStartVals).length > 0) {
+        startVals[key] = nestedStartVals;
+      }
+    }
+  }
+  return startVals;
+}
+
+/**
+ * Get startVals from Query Params
+ *
+ * @param {import("ol/layer").Layer} layer
+ * @param {{[key: string]: any}} layerConfig
+ */
+export function getStartVals(layer, layerConfig) {
+  if (!layer.getSource().getTileUrlFunction()) return null;
+
+  const url = new URL(layer.getSource().getTileUrlFunction()([0, 0, 0]));
+  const queryParams = Object.fromEntries(url.searchParams.entries());
+  const startVals = getNestedStartVals(
+    layerConfig.schema?.properties || layerConfig.schema,
+    queryParams
+  );
+
+  return Object.keys(startVals).length ? startVals : null;
+}