+++ elements/itemfilter/CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## [0.10.1](https://github.com/EOX-A/EOxElements/compare/itemfilter-v0.10.0...itemfilter-v0.10.1) (2023-11-20)
+
+
+### Bug Fixes
+
+* **style:** Max height of scroll container ([5d4b362](https://github.com/EOX-A/EOxElements/commit/5d4b362650b71feee488710d183b944ecad79074))
+
+## [0.10.0](https://github.com/EOX-A/EOxElements/compare/itemfilter-v0.9.0...itemfilter-v0.10.0) (2023-11-16)
+
+
+### Features
+
+* Autocomplete & inline mode ([#422](https://github.com/EOX-A/EOxElements/issues/422)) ([10bdf21](https://github.com/EOX-A/EOxElements/commit/10bdf2192e1dd655fb97ca006b5fa02066fb9de9))
+
+
+### Bug Fixes
+
+* Remove live directive for checking mode matching ([fdd387b](https://github.com/EOX-A/EOxElements/commit/fdd387b8a5f4508d51dcd008f7b143fdb5eb5255))
+
 ## [0.9.0](https://github.com/EOX-A/EOxElements/compare/itemfilter-v0.8.6...itemfilter-v0.9.0) (2023-10-31)